Refactor USD material recreator responsibilities
