+++ src/materials_processor/usd/texture_builders/__init__.py
@@ -1,0 +1,11 @@
+"""Renderer-specific texture material builders."""
+
+from materials_processor.usd.texture_builders.arnold import ArnoldTextureMaterialMixin
+from materials_processor.usd.texture_builders.mtlx import MaterialXTextureMaterialMixin
+from materials_processor.usd.texture_builders.preview import USDPreviewTextureMaterialMixin
+
+__all__ = [
+    "ArnoldTextureMaterialMixin",
+    "MaterialXTextureMaterialMixin",
+    "USDPreviewTextureMaterialMixin",
+]
