+++ src/materials_processor/usd/texture_builders/arnold.py
@@ -1,0 +1,308 @@
+"""Arnold texture material builder helpers."""
+
+import logging
+
+from pxr import Sdf, UsdShade
+
+logger = logging.getLogger(__name__)
+
+
+class ArnoldTextureMaterialMixin:
+    """Create Arnold texture shader networks."""
+
+    ###  arnold ###
+    def _arnold_create_material(self, parent_path, enable_transmission=False):
+        """
+        example prints for variables created by the script:
+            shader: UsdShade.Shader(Usd.Prim(</root/material/mat_hello_world_collect/standard_surface1>))
+            material_prim: Usd.Prim(</root/material/mat_hello_world_collect>)
+            material_usdshade: UsdShade.Material(Usd.Prim(</root/material/mat_hello_world_collect>))
+        """
+        shader_path = f'{parent_path}/arnold_standard_surface1'
+        stdsurf_usdshade = UsdShade.Shader.Define(self.stage, shader_path)
+        stdsurf_usdshade.CreateIdAttr("arnold:standard_surface")
+        material_prim = self.stage.GetPrimAtPath(parent_path)
+
+        material_usdshade = UsdShade.Material.Define(self.stage, material_prim.GetPath())
+        material_usdshade.CreateOutput("arnold:surface", Sdf.ValueTypeNames.Token).ConnectToSource(stdsurf_usdshade.ConnectableAPI(), "surface")
+
+        self._arnold_initialize_standard_surface_shader(stdsurf_usdshade)
+        self._arnold_fill_texture_file_paths(material_prim, stdsurf_usdshade)
+
+        if enable_transmission:
+            self._arnold_enable_transmission(stdsurf_usdshade)
+
+        return material_usdshade
+
+    def _arnold_initialize_standard_surface_shader(self, shader_usdshade):
+        """
+        initializes Arnold Standard Surface inputs
+        """
+        shader_usdshade.CreateInput('aov_id1', Sdf.ValueTypeNames.Float3).Set((0, 0, 0))
+        shader_usdshade.CreateInput('aov_id2', Sdf.ValueTypeNames.Float3).Set((0, 0, 0))
+        shader_usdshade.CreateInput('aov_id3', Sdf.ValueTypeNames.Float3).Set((0, 0, 0))
+        shader_usdshade.CreateInput('aov_id4', Sdf.ValueTypeNames.Float3).Set((0, 0, 0))
+        shader_usdshade.CreateInput('aov_id5', Sdf.ValueTypeNames.Float3).Set((0, 0, 0))
+        shader_usdshade.CreateInput('aov_id6', Sdf.ValueTypeNames.Float3).Set((0, 0, 0))
+        shader_usdshade.CreateInput('aov_id7', Sdf.ValueTypeNames.Float3).Set((0, 0, 0))
+        shader_usdshade.CreateInput('aov_id8', Sdf.ValueTypeNames.Float3).Set((0, 0, 0))
+        shader_usdshade.CreateInput('base', Sdf.ValueTypeNames.Float).Set(1)
+        shader_usdshade.CreateInput('base_color', Sdf.ValueTypeNames.Float3).Set((0.8, 0.8, 0.8))
+        shader_usdshade.CreateInput('metalness', Sdf.ValueTypeNames.Float).Set(0)
+        shader_usdshade.CreateInput('specular', Sdf.ValueTypeNames.Float).Set(1)
+        shader_usdshade.CreateInput('specular_color', Sdf.ValueTypeNames.Float3).Set((1, 1, 1))
+        shader_usdshade.CreateInput('specular_roughness', Sdf.ValueTypeNames.Float).Set(0.2)
+        shader_usdshade.CreateInput('specular_IOR', Sdf.ValueTypeNames.Float).Set(1.5)
+        shader_usdshade.CreateInput('specular_anisotropy', Sdf.ValueTypeNames.Float).Set(0)
+        shader_usdshade.CreateInput('specular_rotation', Sdf.ValueTypeNames.Float).Set(0)
+        shader_usdshade.CreateInput('caustics', Sdf.ValueTypeNames.Bool).Set(False)
+        shader_usdshade.CreateInput('coat', Sdf.ValueTypeNames.Float).Set(0.0)
+        shader_usdshade.CreateInput('coat_color', Sdf.ValueTypeNames.Float3).Set((1, 1, 1))
+        shader_usdshade.CreateInput('coat_roughness', Sdf.ValueTypeNames.Float).Set(0.1)
+        shader_usdshade.CreateInput('coat_IOR', Sdf.ValueTypeNames.Float).Set(1.5)
+        shader_usdshade.CreateInput('coat_normal', Sdf.ValueTypeNames.Float3).Set((0, 0, 0))
+        shader_usdshade.CreateInput('coat_affect_color', Sdf.ValueTypeNames.Float).Set(0)
+        shader_usdshade.CreateInput('coat_affect_roughness', Sdf.ValueTypeNames.Float).Set(0)
+        shader_usdshade.CreateInput('indirect_diffuse', Sdf.ValueTypeNames.Float).Set(1)
+        shader_usdshade.CreateInput('indirect_specular', Sdf.ValueTypeNames.Float).Set(1)
+        shader_usdshade.CreateInput('indirect_reflections', Sdf.ValueTypeNames.Bool).Set(True)
+        shader_usdshade.CreateInput('subsurface', Sdf.ValueTypeNames.Float).Set(0)
+        shader_usdshade.CreateInput('subsurface_anisotropy', Sdf.ValueTypeNames.Float).Set(0)
+        shader_usdshade.CreateInput('subsurface_color', Sdf.ValueTypeNames.Float3).Set((1, 1, 1))
+        shader_usdshade.CreateInput('subsurface_radius', Sdf.ValueTypeNames.Float3).Set((1, 1, 1))
+        shader_usdshade.CreateInput('subsurface_scale', Sdf.ValueTypeNames.Float).Set(1)
+        shader_usdshade.CreateInput('subsurface_type', Sdf.ValueTypeNames.String).Set("randomwalk")
+        shader_usdshade.CreateInput('emission', Sdf.ValueTypeNames.Float).Set(0)
+        shader_usdshade.CreateInput('emission_color', Sdf.ValueTypeNames.Float3).Set((1, 1, 1))
+        shader_usdshade.CreateInput('normal', Sdf.ValueTypeNames.Float3).Set((0, 0, 0))
+        shader_usdshade.CreateInput('opacity', Sdf.ValueTypeNames.Float3).Set((1, 1, 1))
+        shader_usdshade.CreateInput('sheen', Sdf.ValueTypeNames.Float).Set(0)
+        shader_usdshade.CreateInput('sheen_color', Sdf.ValueTypeNames.Float3).Set((1, 1, 1))
+        shader_usdshade.CreateInput('sheen_roughness', Sdf.ValueTypeNames.Float).Set(0.3)
+        shader_usdshade.CreateInput('indirect_diffuse', Sdf.ValueTypeNames.Float).Set(1)
+        shader_usdshade.CreateInput('indirect_specular', Sdf.ValueTypeNames.Float).Set(1)
+        shader_usdshade.CreateInput('internal_reflections', Sdf.ValueTypeNames.Bool).Set(True)
+        shader_usdshade.CreateInput('caustics', Sdf.ValueTypeNames.Bool).Set(False)
+        shader_usdshade.CreateInput('exit_to_background', Sdf.ValueTypeNames.Bool).Set(False)
+        shader_usdshade.CreateInput('tangent', Sdf.ValueTypeNames.Float3).Set((0, 0, 0))
+        shader_usdshade.CreateInput('transmission', Sdf.ValueTypeNames.Float).Set(0)
+        shader_usdshade.CreateInput('transmission_color', Sdf.ValueTypeNames.Float3).Set((1, 1, 1))
+        shader_usdshade.CreateInput('transmission_depth', Sdf.ValueTypeNames.Float).Set(0)
+        shader_usdshade.CreateInput('transmission_scatter', Sdf.ValueTypeNames.Float3).Set((0, 0, 0))
+        shader_usdshade.CreateInput('transmission_scatter_anisotropy', Sdf.ValueTypeNames.Float).Set(0)
+        shader_usdshade.CreateInput('transmission_dispersion', Sdf.ValueTypeNames.Float).Set(0)
+        shader_usdshade.CreateInput('transmission_extra_roughness', Sdf.ValueTypeNames.Float).Set(0)
+        shader_usdshade.CreateInput('thin_film_IOR', Sdf.ValueTypeNames.Float).Set(1.5)
+        shader_usdshade.CreateInput('thin_film_thickness', Sdf.ValueTypeNames.Float).Set(0)
+        shader_usdshade.CreateInput('thin_walled', Sdf.ValueTypeNames.Bool).Set(False)
+        shader_usdshade.CreateInput('transmit_aovs', Sdf.ValueTypeNames.Bool).Set(False)
+
+    def _arnold_initialize_image_shader(self, image_path: str):
+        image_shader = UsdShade.Shader.Define(self.stage, image_path)
+        image_shader.CreateIdAttr("arnold:image")
+
+        color_space = image_shader.CreateInput("color_space", Sdf.ValueTypeNames.String)
+        color_space.Set("auto")
+        file_input = image_shader.CreateInput("filename", Sdf.ValueTypeNames.Asset)
+        filter = image_shader.CreateInput("filter", Sdf.ValueTypeNames.String)
+        filter.Set("smart_bicubic")
+        ignore_missing_textures = image_shader.CreateInput("ignore_missing_textures", Sdf.ValueTypeNames.Bool)
+        ignore_missing_textures.Set(False)
+        mipmap_bias = image_shader.CreateInput("mipmap_bias", Sdf.ValueTypeNames.Int)
+        mipmap_bias.Set(0)
+        missing_texture_color = image_shader.CreateInput("missing_texture_color", Sdf.ValueTypeNames.Float4)
+        missing_texture_color.Set((0,0,0,0))
+        multiply = image_shader.CreateInput("multiply", Sdf.ValueTypeNames.Float3)
+        multiply.Set((1,1,1))
+        offset = image_shader.CreateInput("offset", Sdf.ValueTypeNames.Float3)
+        offset.Set((0,0,0))
+        sflip = image_shader.CreateInput("sflip", Sdf.ValueTypeNames.Bool)
+        sflip.Set(False)
+        single_channel = image_shader.CreateInput("single_channel", Sdf.ValueTypeNames.Bool)
+        single_channel.Set(False)
+        soffset = image_shader.CreateInput("soffset", Sdf.ValueTypeNames.Float)
+        soffset.Set(0)
+        sscale = image_shader.CreateInput("sscale", Sdf.ValueTypeNames.Float)
+        sscale.Set(1)
+        start_channel = image_shader.CreateInput("start_channel", Sdf.ValueTypeNames.Int)
+        start_channel.Set(0)
+        swap_st = image_shader.CreateInput("swap_st", Sdf.ValueTypeNames.Bool)
+        swap_st.Set(False)
+        swrap = image_shader.CreateInput("swrap", Sdf.ValueTypeNames.String)
+        swrap.Set("periodic")
+        tflip = image_shader.CreateInput("tflip", Sdf.ValueTypeNames.Bool)
+        tflip.Set(False)
+        toffset = image_shader.CreateInput("toffset", Sdf.ValueTypeNames.Float)
+        toffset.Set(0)
+        tscale = image_shader.CreateInput("tscale", Sdf.ValueTypeNames.Float)
+        tscale.Set(1)
+        twrap = image_shader.CreateInput("twrap", Sdf.ValueTypeNames.String)
+        twrap.Set("periodic")
+        uvcoords = image_shader.CreateInput("uvcoords", Sdf.ValueTypeNames.Float2)
+        uvcoords.Set((0,0))
+        uvset = image_shader.CreateInput("uvset", Sdf.ValueTypeNames.String)
+        uvset.Set("")
+
+        return image_shader
+
+    def _arnold_initialize_color_correct_shader(self, color_correct_path: str):
+        color_correct_shader = UsdShade.Shader.Define(self.stage, color_correct_path)
+        color_correct_shader.CreateIdAttr("arnold:color_correct")
+        cc_add_input = color_correct_shader.CreateInput("add", Sdf.ValueTypeNames.Float3)
+        cc_add_input.Set((0, 0, 0))
+        cc_contrast_input = color_correct_shader.CreateInput("contrast", Sdf.ValueTypeNames.Float)
+        cc_contrast_input.Set(1)
+        cc_exposure_input = color_correct_shader.CreateInput("exposure", Sdf.ValueTypeNames.Float)
+        cc_exposure_input.Set(0)
+        cc_gamma_input = color_correct_shader.CreateInput("gamma", Sdf.ValueTypeNames.Float)
+        cc_gamma_input.Set(1)
+        cc_hue_shift_input = color_correct_shader.CreateInput("hue_shift", Sdf.ValueTypeNames.Float)
+        cc_hue_shift_input.Set(0)
+
+        return color_correct_shader
+
+    def _arnold_initialize_range_shader(self, range_path: str):
+        range_shader = UsdShade.Shader.Define(self.stage, range_path)
+        range_shader.CreateIdAttr("arnold:range")
+
+        bias_input = range_shader.CreateInput("bias", Sdf.ValueTypeNames.Float)
+        bias_input.Set(0.5)
+        contrast_input = range_shader.CreateInput("contrast", Sdf.ValueTypeNames.Float)
+        contrast_input.Set(1)
+        contrast_pivot_input = range_shader.CreateInput("contrast_pivot", Sdf.ValueTypeNames.Float)
+        contrast_pivot_input.Set(0.5)
+        gain_input = range_shader.CreateInput("gain", Sdf.ValueTypeNames.Float)
+        gain_input.Set(0.5)
+        input_min_input = range_shader.CreateInput("input_min", Sdf.ValueTypeNames.Float)
+        input_min_input.Set(0)
+        input_max_input = range_shader.CreateInput("input_max", Sdf.ValueTypeNames.Float)
+        input_max_input.Set(1)
+        output_min_input = range_shader.CreateInput("output_min", Sdf.ValueTypeNames.Float)
+        output_min_input.Set(0)
+        output_max_input = range_shader.CreateInput("output_max", Sdf.ValueTypeNames.Float)
+        output_max_input.Set(1)
+        output_max_input = range_shader.CreateInput("smoothstep", Sdf.ValueTypeNames.Bool)
+        output_max_input.Set(False)
+
+        return range_shader
+
+
+    def _arnold_initialize_normal_map_shader(self, normal_map_path: str):
+        normal_map_shader = UsdShade.Shader.Define(self.stage, normal_map_path)
+        normal_map_shader.CreateIdAttr("arnold:normal_map")
+
+        color_to_signed_input = normal_map_shader.CreateInput("color_to_signed", Sdf.ValueTypeNames.Bool)
+        color_to_signed_input.Set(True)
+        input_input = normal_map_shader.CreateInput("input", Sdf.ValueTypeNames.Float3)
+        input_input.Set((0, 0, 0))
+        invert_x_input = normal_map_shader.CreateInput("invert_x", Sdf.ValueTypeNames.Bool)
+        invert_x_input.Set(False)
+        invert_y_input = normal_map_shader.CreateInput("invert_y", Sdf.ValueTypeNames.Bool)
+        invert_y_input.Set(False)
+        invert_z_input = normal_map_shader.CreateInput("invert_z", Sdf.ValueTypeNames.Bool)
+        invert_z_input.Set(False)
+        normal_input = normal_map_shader.CreateInput("normal", Sdf.ValueTypeNames.Float3)
+        normal_input.Set((0, 0, 0))
+        order_input = normal_map_shader.CreateInput("order", Sdf.ValueTypeNames.String)
+        order_input.Set('XYZ')
+        strength_input = normal_map_shader.CreateInput("strength", Sdf.ValueTypeNames.Float)
+        strength_input.Set(1)
+        tangent_input = normal_map_shader.CreateInput("tangent", Sdf.ValueTypeNames.Float3)
+        tangent_input.Set((0, 0, 0))
+        tangent_space_input = normal_map_shader.CreateInput("tangent_space", Sdf.ValueTypeNames.Bool)
+        tangent_space_input.Set(True)
+
+        return normal_map_shader
+
+    def _arnold_initialize_bump2d_shader(self, bump2d_path: str):
+        bump2d_shader = UsdShade.Shader.Define(self.stage, bump2d_path)
+        bump2d_shader.CreateIdAttr("arnold:bump2d")
+
+        bump_height_input = bump2d_shader.CreateInput("bump_height", Sdf.ValueTypeNames.Float)
+        bump_height_input.Set(1)
+        bump_map_input = bump2d_shader.CreateInput("bump_map", Sdf.ValueTypeNames.Float)
+        bump_map_input.Set(0)
+        normal_input = bump2d_shader.CreateInput("normal", Sdf.ValueTypeNames.Float3)
+        normal_input.Set((0, 0, 0))
+
+        return bump2d_shader
+
+
+    def _arnold_enable_transmission(self, shader_usdshade):
+        """
+        given the mtlx standard surface, will set input primvar 'transmission' to value '0.9'
+        """
+        shader_usdshade.GetInput('transmission').Set(0.9)
+        shader_usdshade.GetInput('thin_walled').Set(True)
+
+
+    def _arnold_fill_texture_file_paths(self, material_prim, std_surf_shader):
+        """
+        Fills the texture file paths for the given shader using the material_data.
+        """
+        # map of tex_type to it's name on an Arnold Standard Surface shader.
+        texture_types_to_inputs = {
+            'basecolor': 'base_color',
+            'metalness': 'metalness',
+            'roughness': 'specular_roughness',
+            'normal': 'normal',
+            'opacity': 'opacity',
+            'height': 'height',
+        }
+
+        bump2d_path = f"{material_prim.GetPath()}/arnold_Bump2d"
+        bump2d_shader = None
+
+        for tex_type, tex_dict in self.material_dict.items():
+            tex_filepath = tex_dict['path']
+            tex_type = tex_type.lower()  # assume all lowercase
+            if tex_type not in texture_types_to_inputs:
+                logger.warning("tex_type: '%s' not supported yet for arnold", tex_type)
+                continue
+
+            input_name = texture_types_to_inputs[tex_type]
+
+            # create arnold::image prim
+            texture_prim_path = f'{material_prim.GetPath()}/arnold_{tex_type}Texture'
+            texture_shader = self._arnold_initialize_image_shader(texture_prim_path)
+            texture_shader.GetInput("filename").Set(tex_filepath)
+
+            if tex_type in ['basecolor']:
+                color_correct_path = f"{material_prim.GetPath()}/arnold_{tex_type}ColorCorrect"
+                color_correct_shader = self._arnold_initialize_color_correct_shader(color_correct_path)
+                color_correct_shader.CreateInput("input", Sdf.ValueTypeNames.Float4).ConnectToSource(texture_shader.ConnectableAPI(), "rgba")
+                std_surf_shader.CreateInput(input_name, Sdf.ValueTypeNames.Float3).ConnectToSource(color_correct_shader.ConnectableAPI(), "rgb")
+
+            elif tex_type in ['metalness']:
+                # disable metalness if material is transmissive like glass:
+                if self.is_transmissive:
+                    continue
+                range_path = f"{material_prim.GetPath()}/arnold_{tex_type}Range"
+                range_shader = self._arnold_initialize_range_shader(range_path)
+                range_shader.CreateInput("input", Sdf.ValueTypeNames.Float4).ConnectToSource(texture_shader.ConnectableAPI(), "rgba")
+                std_surf_shader.CreateInput(input_name, Sdf.ValueTypeNames.Float3).ConnectToSource(range_shader.ConnectableAPI(), "r")
+
+            elif tex_type in ['roughness']:
+                range_path = f"{material_prim.GetPath()}/arnold_{tex_type}Range"
+                range_shader = self._arnold_initialize_range_shader(range_path)
+                range_shader.CreateInput("input", Sdf.ValueTypeNames.Float4).ConnectToSource(texture_shader.ConnectableAPI(), "rgba")
+                std_surf_shader.CreateInput(input_name, Sdf.ValueTypeNames.Float3).ConnectToSource(range_shader.ConnectableAPI(), "r")
+
+            elif tex_type in ['height']:
+                range_path = f"{material_prim.GetPath()}/arnold_{tex_type}Range"
+                range_shader = self._arnold_initialize_range_shader(range_path)
+                range_shader.CreateInput("input", Sdf.ValueTypeNames.Float4).ConnectToSource(texture_shader.ConnectableAPI(), "rgba")
+                if not bump2d_shader:
+                    bump2d_shader = self._arnold_initialize_bump2d_shader(bump2d_path)
+                bump2d_shader.CreateInput("bump_map", Sdf.ValueTypeNames.Float).ConnectToSource(range_shader.ConnectableAPI(), "r")
+
+            elif tex_type in ['normal']:
+                normal_map_path = f"{material_prim.GetPath()}/arnold_NormalMap"
+                normal_map_shader = self._arnold_initialize_normal_map_shader(normal_map_path)
+                normal_map_shader.CreateInput("input", Sdf.ValueTypeNames.Float3).ConnectToSource(texture_shader.ConnectableAPI(), "vector")
+                if not bump2d_shader:
+                    bump2d_shader = self._arnold_initialize_bump2d_shader(bump2d_path)
+                bump2d_shader.CreateInput("normal", Sdf.ValueTypeNames.Float4).ConnectToSource(normal_map_shader.ConnectableAPI(), "vector")
+
+        if bump2d_shader:
+            std_surf_shader.CreateInput('normal', Sdf.ValueTypeNames.Float3).ConnectToSource(bump2d_shader.ConnectableAPI(), "vector")
